[chore](refactor-params)remove old properties (#56163)
remove oldcode

(cherry picked from commit 87f8248962f0e32353e47985c45e21d18dc33e0b)
