+++ fe/fe-core/src/main/java/org/apache/doris/analysis/CreateResourceStmt.java
@@ -25,8 +25,7 @@ import org.apache.doris.common.ErrorCode;
 import org.apache.doris.common.ErrorReport;
 import org.apache.doris.common.FeNameFormat;
 import org.apache.doris.common.UserException;
-import org.apache.doris.common.util.PrintableMap;
-import org.apache.doris.datasource.property.constants.AzureProperties;
+import org.apache.doris.datasource.property.storage.AzureProperties;
 import org.apache.doris.mysql.privilege.PrivPredicate;
 import org.apache.doris.qe.ConnectContext;
 
@@ -82,7 +81,7 @@ public class CreateResourceStmt extends DdlStmt implements NotFallbackInParser {
             throw new AnalysisException("Resource type can't be null");
         }
 
-        if (AzureProperties.checkAzureProviderPropertyExist(properties)) {
+        if (AzureProperties.guessIsMe(properties)) {
             resourceType = ResourceType.AZURE;
             return;
         }
