+++ fe/fe-core/src/main/java/org/apache/doris/catalog/HdfsStorageVault.java
@@ -22,7 +22,6 @@ import org.apache.doris.cloud.proto.Cloud;
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.security.authentication.AuthenticationConfig;
 import org.apache.doris.common.util.PrintableMap;
-import org.apache.doris.datasource.property.constants.S3Properties;
 import org.apache.doris.datasource.property.storage.HdfsCompatibleProperties;
 import org.apache.doris.datasource.property.storage.StorageProperties;
 import org.apache.doris.fs.remote.dfs.DFSFileSystem;
@@ -59,7 +58,7 @@ import java.util.stream.Collectors;
  */
 public class HdfsStorageVault extends StorageVault {
     private static final Logger LOG = LogManager.getLogger(HdfsStorageVault.class);
-
+    public static final String S3_VALIDITY_CHECK = "s3_validity_check";
     public static final String HADOOP_FS_PREFIX = "dfs.";
     public static String HADOOP_SHORT_CIRCUIT = "dfs.client.read.shortcircuit";
     public static String HADOOP_SOCKET_PATH = "dfs.domain.socket.path";
@@ -87,7 +86,7 @@ public class HdfsStorageVault extends StorageVault {
      * such as `type`, `path_prefix`, etc.
      */
     private static final Set<String> NON_HDFS_CONF_PROPERTY_KEYS =
-            ImmutableSet.of(StorageVault.PropertyKey.TYPE, PropertyKey.VAULT_PATH_PREFIX, S3Properties.VALIDITY_CHECK)
+            ImmutableSet.of(StorageVault.PropertyKey.TYPE, PropertyKey.VAULT_PATH_PREFIX, S3_VALIDITY_CHECK)
                     .stream().map(String::toLowerCase)
                     .collect(ImmutableSet.toImmutableSet());
 
@@ -113,8 +112,8 @@ public class HdfsStorageVault extends StorageVault {
     }
 
     public static void checkConnectivity(Map<String, String> newProperties) throws DdlException {
-        if (newProperties.containsKey(S3Properties.VALIDITY_CHECK)
-                && newProperties.get(S3Properties.VALIDITY_CHECK).equalsIgnoreCase("false")) {
+        if (newProperties.containsKey(S3_VALIDITY_CHECK)
+                && newProperties.get(S3_VALIDITY_CHECK).equalsIgnoreCase("false")) {
             return;
         }
