+++ fe/fe-core/src/main/java/org/apache/doris/catalog/S3StorageVault.java
@@ -20,7 +20,8 @@ package org.apache.doris.catalog;
 import org.apache.doris.analysis.CreateResourceStmt;
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.UserException;
-import org.apache.doris.datasource.property.constants.S3Properties;
+import org.apache.doris.datasource.property.storage.S3Properties;
+import org.apache.doris.datasource.property.storage.StorageProperties;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
@@ -58,14 +59,14 @@ import java.util.Map;
 public class S3StorageVault extends StorageVault {
     // Reuse all the code from S3Resource
     private Resource resource;
-
+    public static final String S3_ROOT_PATH = "s3.root.path";
 
     public static class PropertyKey {
         public static final String ACCESS_KEY = S3Properties.ACCESS_KEY;
         public static final String SECRET_KEY = S3Properties.SECRET_KEY;
         public static final String USE_PATH_STYLE = S3Properties.USE_PATH_STYLE;
-        public static final String ROOT_PATH = S3Properties.ROOT_PATH;
-        public static final String PROVIDER = S3Properties.PROVIDER;
+        public static final String ROOT_PATH = S3_ROOT_PATH;
+        public static final String PROVIDER = StorageProperties.FS_PROVIDER_KEY;
         public static final String REGION = S3Properties.REGION;
         public static final String ENDPOINT = S3Properties.ENDPOINT;
         public static final String BUCKET = S3Properties.BUCKET;
