+++ fe/fe-core/src/main/java/org/apache/doris/datasource/lakesoul/source/LakeSoulScanNode.java
@@ -25,8 +25,6 @@ import org.apache.doris.datasource.FileQueryScanNode;
 import org.apache.doris.datasource.TableFormatType;
 import org.apache.doris.datasource.lakesoul.LakeSoulExternalTable;
 import org.apache.doris.datasource.lakesoul.LakeSoulUtils;
-import org.apache.doris.datasource.property.constants.MinioProperties;
-import org.apache.doris.datasource.property.constants.S3Properties;
 import org.apache.doris.planner.PlanNodeId;
 import org.apache.doris.qe.SessionVariable;
 import org.apache.doris.spi.Split;
@@ -181,8 +179,8 @@ public class LakeSoulScanNode extends FileQueryScanNode {
             LOG.debug("{}", catalogProps);
         }
 
-        if (catalogProps.get(S3Properties.Env.ENDPOINT) != null) {
-            options.put(LakeSoulUtils.FS_S3A_ENDPOINT, catalogProps.get(S3Properties.Env.ENDPOINT));
+        if (catalogProps.get("AWS_ENDPOINT") != null) {
+            options.put(LakeSoulUtils.FS_S3A_ENDPOINT, catalogProps.get("AWS_ENDPOINT"));
             if (!options.containsKey("oss.endpoint")) {
                 // Aliyun OSS requires virtual host style access
                 options.put(LakeSoulUtils.FS_S3A_PATH_STYLE_ACCESS, "false");
@@ -190,14 +188,14 @@ public class LakeSoulScanNode extends FileQueryScanNode {
                 // use path style access for all other s3 compatible storage services
                 options.put(LakeSoulUtils.FS_S3A_PATH_STYLE_ACCESS, "true");
             }
-            if (catalogProps.get(S3Properties.Env.ACCESS_KEY) != null) {
-                options.put(LakeSoulUtils.FS_S3A_ACCESS_KEY, catalogProps.get(S3Properties.Env.ACCESS_KEY));
+            if (catalogProps.get("AWS_ACCESS_KEY") != null) {
+                options.put(LakeSoulUtils.FS_S3A_ACCESS_KEY, catalogProps.get("AWS_ACCESS_KEY"));
             }
-            if (catalogProps.get(S3Properties.Env.SECRET_KEY) != null) {
-                options.put(LakeSoulUtils.FS_S3A_SECRET_KEY, catalogProps.get(S3Properties.Env.SECRET_KEY));
+            if (catalogProps.get("AWS_SECRET_KEY") != null) {
+                options.put(LakeSoulUtils.FS_S3A_SECRET_KEY, catalogProps.get("AWS_SECRET_KEY"));
             }
-            if (catalogProps.get(S3Properties.Env.REGION) != null) {
-                options.put(LakeSoulUtils.FS_S3A_REGION, catalogProps.get(S3Properties.Env.REGION));
+            if (catalogProps.get("AWS_REGION") != null) {
+                options.put(LakeSoulUtils.FS_S3A_REGION, catalogProps.get("AWS_REGION"));
             }
         }
