+++ fe/fe-core/src/main/java/org/apache/doris/datasource/property/storage/AzureProperties.java
@@ -94,7 +94,6 @@ public class AzureProperties extends StorageProperties {
     @Getter
     protected String forceParsingByStandardUrl = "false";
 
-
     public AzureProperties(Map<String, String> origProps) {
         super(Type.AZURE, origProps);
     }
@@ -109,7 +108,7 @@ public class AzureProperties extends StorageProperties {
             throw new IllegalArgumentException(String.format("Endpoint '%s' is not valid. It should end with '%s'.",
                     endpoint, AZURE_ENDPOINT_SUFFIX));
         }
-        this.endpoint = formatAzureEndpoint(endpoint);
+        this.endpoint = formatAzureEndpoint(endpoint, accessKey);
     }
 
     public static boolean guessIsMe(Map<String, String> origProps) {
@@ -144,7 +143,7 @@ public class AzureProperties extends StorageProperties {
 
     public static final String AZURE_ENDPOINT_TEMPLATE = "https://%s.blob.core.windows.net";
 
-    private String formatAzureEndpoint(String endpoint) {
+    public static String formatAzureEndpoint(String endpoint, String accessKey) {
         if (Config.force_azure_blob_global_endpoint) {
             return String.format(AZURE_ENDPOINT_TEMPLATE, accessKey);
         }
